Tree: move viewer to visualizations
