+++ Orange/widgets/visualize/owtreeviewer2d.py
